Update v0.8
Changed the way the gestures work (now they can be enabled and disable, did this for the reason that the Menu, Tutorial and Playtime will be all in the same scene) this should evade some errors as well. Because the gloves broke I had to adjust some stuff so I couldnt finish the tutorial screen unfortunately. Will do it later on this weekend.